Update for commented tidy code
